+++ routes/movies.js
@@ -21,25 +21,25 @@ const validateURL = (value) => {
   return value;
 };
 
-router.get('/', getMovies);
-router.post('/', celebrate({
+router.get('/movies', getMovies);
+router.post('/movies', celebrate({
   body: Joi.object().keys({
     country: Joi.string().min(3).max(30).required(),
     director: Joi.string().min(3).max(30).required(),
     duration: Joi.number().integer().required(),
     year: Joi.number().integer().required(),
-    description: Joi.string().min(3).max(120).required(),
+    description: Joi.string().min(3).max(1000).required(),
     image: Joi.string().custom(validateURL).required(),
     trailerLink: Joi.string().custom(validateURL).required(),
     thumbnail: Joi.string().custom(validateURL).required(),
     nameRU: Joi.string().min(3).max(30).required(),
     nameEN: Joi.string().min(3).max(30).required(),
-    movieId: Joi.string().length(24).hex(),
+    movieId: Joi.number().integer().required(),
   }),
 }), createMovie);
-router.delete('/:movieId', celebrate({
-  body: Joi.object().keys({
-    movieId: Joi.string().length(24).hex(),
+router.delete('/movies/:movieId', celebrate({
+  params: Joi.object().keys({
+    movieId: Joi.string().hex().length(24).required(),
   }),
 }), delMovieById);
